No longer override Django's native timeout parameter handling
Django has since fixed the issue that prevented setting an infinite
timeout, which can now be done using a timeout value of `None`. However
their chosen solution is not consistent with the approach taken by
django-pylibmc's custom `get_backend_timeout()`.

Even though this is a non-backwards compatible change for people using a
timeout value of `0` with django-pylibmc, the custom handling should
still be removed, to:
* simplify the django-pylibmc codebase
* make django-pylibmc's behaviour consistent with the Django docs
* avoid changes in behaviour when people switch from the native backend
  to django-pylibmc, expecting it to be a drop-in replacement.

The adjusted test was copied from:
https://github.com/django/django/blob/a6d463b0965bcdabc202450f00d2c09f944f8edb/tests/cache/tests.py#L522-L534

This change will need a major version bump of django-pylibmc.

Fixes #35.