Stop closing pylibmc connections after each request
Since this resets the failover state, and creates an unnecessary
amount of reconnects. This is the same fix that was applied to Django's
own pylibmc backend in django/django#7200.

(django-pylibmc inherits from `BaseMemcachedCache` so doesn't benefit
from the upstream fix - in the future that should probably change,
however it will be a more invasive fix, due to it also affecting the
behaviour of `_cache()`)

This fix reduced our average query time in production by 70%.

Fixes #6.